+++ tests/plugins/xcube/test_plugin.py
@@ -32,9 +32,9 @@ class ExportPluginTest(TestCase):
         all_rule_names = set(f"xcube/{k}" for k in plugin.rules.keys())
         self.assertEqual(
             all_rule_names,
-            set(plugin.configs["all"].configs[-1].rules.keys()),
+            set(plugin.configs["all"][-1].rules.keys()),
         )
         self.assertEqual(
             all_rule_names,
-            set(plugin.configs["recommended"].configs[-1].rules.keys()),
+            set(plugin.configs["recommended"][-1].rules.keys()),
         )